fix: tamaño del nav-link & header para mobile
